requirements.txt
Django==4.2
Pillow
psycopg2-binary
python-dotenv
flake8
django-debug-toolbar
